Class update

+++ mysqlconnector.py
@@ -4,13 +4,16 @@
 
 import mysql.connector
 import config # project config file
-
+import markets
+import logging
 
 class MySqlExchangeProcessor():
     """ Used to establish connection to MySql database, prepare clean table structure based on markets data
     and perform insertion and updates of information based on messages from exchanges"""
 
     def __init__(self, user=None, password=None, host='127.0.0.1', database=None):
+        self.database_ready = False
+
         if database:
             self._database = database
         # Try to connect to DB
@@ -20,13 +23,14 @@ class MySqlExchangeProcessor():
             print("Error %d: %s" % (e.args[0], e.args[1]))
             sys.exit(1)
 
-
-
         try:
-            self.prepare_database()
+            database_ready = self.prepare_database()
         except mysql.connector.Error as e:
             print("Error %d: %s" % (e.args[0], e.args[1]))
             sys.exit(1)
+
+        if not database_ready:
+            raise RuntimeError
 
     def prepare_database(self):
         # Get list of available tables and exclude items with "_template" ending
@@ -34,24 +38,36 @@ class MySqlExchangeProcessor():
         for x in self.get_tables():
             if x[0] and "_template" not in x[0]:
                 _list_of_tables.append(x[0])
-        print(_list_of_tables)
 
-        # Prepare drop statements
-        _sql_tables_drop_stmt = "DROP TABLE IF EXISTS %s" % ', '.join(_list_of_tables)
-        print(_sql_tables_drop_stmt)
+        # Prepare and execute drop statements (if tables exist)
+        if len(_list_of_tables) > 0:
+            _sql_tables_drop_stmt = "DROP TABLE IF EXISTS %s" % ', '.join(_list_of_tables)
 
+            _cursor = self._connection.cursor()
+            print('Dropping all existing tables (excluding templates)...')
+            try:
+                _cursor.execute(_sql_tables_drop_stmt)
+                _cursor.close()
+            except ValueError as e:
+                print(e)
+
+        # Prepare and execute create table statements for all markets
+        _sql_tables_create_stmt = []
+        for each in markets.list_all_markets():
+            _sql_tables_create_stmt.append("CREATE TABLE %s_sell_orders LIKE _sell_order_template" % each)
+            # _sql_tables_create_stmt += "CREATE TABLE %s_buy_orders LIKE __buy_order_template;\n" % each
+            # _sql_tables_create_stmt += "CREATE TABLE %s_trades LIKE _trade_template;\n" % each
         _cursor = self._connection.cursor()
+        print('Creating new tables for all markets from templates...')
         try:
-            _cursor.execute(_sql_tables_drop_stmt)
-        except ValueError:
-            pass
-        # Drop all tables from updated list
+            for each in _sql_tables_create_stmt:
+                _cursor.execute(each)
+            _cursor.close()
+        except mysql.connector.Error as e:
+            print(e)
 
-
-        # Create new tables based on markets['byCurrencyPair'] using templates
-
-
-
+        # Database initialized
+        return True
 
     def close(self):
         self._connection.close()
@@ -88,20 +104,19 @@ class MySqlExchangeProcessor():
             print(_q)
 
 
-    def init_tables_for_markets(self):
-        pass
 
-mysql = MySqlExchangeProcessor(
-                                user=config.config['db_user'],
-                                password=config.config['db_pass'],
-                                host=config.config['db_host'],
-                                database=config.config['db_name']
-)
-
+if __name__ == "__main__":
+    mysql = MySqlExchangeProcessor(
+                                    user=config.config['db_user'],
+                                    password=config.config['db_pass'],
+                                    host=config.config['db_host'],
+                                    database=config.config['db_name']
+    )
 
 
 
-mysql.close()
+
+    mysql.close()
